+++ CHANGELOG.md
@@ -1,3 +1,4 @@
 > Read detailed change logs in https://github.com/sakurawald/fuji/commits/dev/
 
+- [jail] feature: new `/jail gui` command, to open a GUI to inspect all defined jails and their jail records.
 - [economy & leaderboard] feature: improve the compatibility with async read/write operations.